chore: remove unused TypeScript configuration files and update import paths for CardImage and CardSkeleton components
