+++ src/lib/components/card-skeleton/card-skeleton.tsx
@@ -1,0 +1,19 @@
+import {
+  BodySkeleton,
+  CardSkeletonContainer,
+  FooterSkeleton,
+  TitleSkeleton,
+} from './card-skeleton.styled';
+import { ImageSkeleton } from './card-skeleton.styled';
+
+export const CardSkeleton = () => {
+  return (
+    <CardSkeletonContainer>
+      <ImageSkeleton />
+      <BodySkeleton>
+        <TitleSkeleton />
+      </BodySkeleton>
+      <FooterSkeleton />
+    </CardSkeletonContainer>
+  );
+};
