+++ src/lib/components/skeleton-cards-list/cards-skeleton-list.tsx
@@ -1,5 +1,6 @@
 import styled from 'styled-components';
-import { CardSkeleton } from '../card/card-skeleton';
+import { CardSkeleton } from '../card-skeleton/card-skeleton';
+
 
 const SkeletonContainer = styled.div`
   padding: 2rem;
